Refactor: Move createAttendanceManagement method from AttendanceManageMent to MyAttendance
