+++ src/main/java/kr/happytravel/erp/attendances/controller/MyAttendanceController.java
@@ -1,5 +1,6 @@
 package kr.happytravel.erp.attendances.controller;
 
+import kr.happytravel.erp.attendances.model.AttendanceManagementModel;
 import kr.happytravel.erp.attendances.model.MyAttendanceResponseModel;
 import kr.happytravel.erp.attendances.model.MyVacationResponseModel;
 import kr.happytravel.erp.attendances.service.MyAttendanceService;
@@ -7,11 +8,11 @@ import lombok.RequiredArgsConstructor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @RequestMapping("/attendance")
@@ -48,6 +49,24 @@ public class MyAttendanceController {
             List<MyVacationResponseModel> myVacationList = myAttendanceService.getMyVacationList(empId);
             logger.info("Detched " + myVacationList + " empList.");
             return ResponseEntity.ok(myVacationList);
+        } catch (IllegalArgumentException e) {
+            logger.warn("Invalid argument: " + e.getMessage());
+            throw e;
+        } catch (Exception e) {
+            logger.error("An error occurred: " + e.getMessage(), e);
+            throw e;
+        }
+    }
+
+    // Create
+    @PostMapping("/add")
+    public ResponseEntity<String> createAttendanceManagement(@RequestBody AttendanceManagementModel attendanceManagement, HttpServletRequest request,
+                                                             HttpServletResponse response, HttpSession session) throws Exception {
+        try {
+            logger.info("Received request to create attendanceManagement: " + attendanceManagement);
+            int result = myAttendanceService.insertAttendanceManagement(attendanceManagement);
+            logger.info("Created attendance, result: " + result);
+            return ResponseEntity.ok("AttendanceManagement created successfully");
         } catch (IllegalArgumentException e) {
             logger.warn("Invalid argument: " + e.getMessage());
             throw e;
