+++ src/main/java/kr/happytravel/erp/attendances/dao/MyAttendanceDao.java
@@ -1,5 +1,6 @@
 package kr.happytravel.erp.attendances.dao;
 
+import kr.happytravel.erp.attendances.model.AttendanceManagementModel;
 import kr.happytravel.erp.attendances.model.MyAttendanceResponseModel;
 import kr.happytravel.erp.attendances.model.MyVacationResponseModel;
 
@@ -8,4 +9,5 @@ import java.util.List;
 public interface MyAttendanceDao {
     List<MyAttendanceResponseModel> getMyAttendanceList(String empId);
     List<MyVacationResponseModel> getMyVacationList(String empId);
+    int insertAttendanceManagement(AttendanceManagementModel attendanceManagement) throws Exception;
 }
