+++ src/main/java/kr/happytravel/erp/attendances/service/MyAttendanceServiceImpl.java
@@ -1,6 +1,7 @@
 package kr.happytravel.erp.attendances.service;
 
 import kr.happytravel.erp.attendances.dao.MyAttendanceDao;
+import kr.happytravel.erp.attendances.model.AttendanceManagementModel;
 import kr.happytravel.erp.attendances.model.MyAttendanceResponseModel;
 import kr.happytravel.erp.attendances.model.MyVacationResponseModel;
 import lombok.RequiredArgsConstructor;
@@ -21,6 +22,11 @@ public class MyAttendanceServiceImpl implements MyAttendanceService {
     @Override
     public List<MyVacationResponseModel> getMyVacationList(String empId) {
         return myAttendanceDao.getMyVacationList(empId);
+    }
+
+    @Override
+    public int insertAttendanceManagement(AttendanceManagementModel attendanceManagement) throws Exception {
+        return myAttendanceDao.insertAttendanceManagement(attendanceManagement);
     }
 
 }
